requirements.txt
selenium
google-cloud-storage